ch54

+++ os/src/config.rs
@@ -25,4 +25,4 @@ pub const CLOCK_FREQ: usize = 12500000;
 pub const MEMORY_END: usize = 0x88000000;
 
 /// stride number
-pub const BIG_STRIDE: usize = 1 << 30;
+pub const BIG_STRIDE: u32 = 1 << 30;
